--- src/ca/mss/impl/multitest/NewThread.java
@@ -1,14 +1,0 @@
-package ca.mss.impl.multitest;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface NewThread {
-	
-	public boolean runLast() default false;
-
-}
