+++ src/ca/mss/multitest/NewThread.java
@@ -1,0 +1,14 @@
+package ca.mss.multitest;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface NewThread {
+	
+	public boolean runLast() default false;
+
+}
