+++ test/ca/mss/test/multitest/increment/Increment_SingleThread_Atomic.java
@@ -2,9 +2,9 @@ package ca.mss.test.multitest.increment;
 
 import org.junit.Test;
 
-import ca.mss.impl.multitest.ConcurrentTest;
-import ca.mss.impl.multitest.MultiTest;
-import ca.mss.impl.multitest.NewThread;
+import ca.mss.multitest.ConcurrentTest;
+import ca.mss.multitest.MultiTest;
+import ca.mss.multitest.NewThread;
 
 @ConcurrentTest( maxTry = IncrementSuite.MAX_TRY )
 public class Increment_SingleThread_Atomic extends Increment {
